backend/otahub/requirements.txt
django
djangorestframework
pillow
django-cors-headers
djangorestframework-simplejwt
stripe
python-dotenv
